\name{RemoteREngine.examples-package}
\alias{RemoteREngine.examples-package}
\alias{RemoteREngine.examples}
\docType{package}
\title{
Set of examples for the RemoteREngine package
}
\description{
This package contains several small java projects 
acting as clients for the RemoteREngine server
}
\details{
\tabular{ll}{
Package: \tab RemoteREngine.examples\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-09-15\cr
License: \tab GPL (>= 3.0)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Rmain Francois <francoisromain@free.fr>

Maintainer: Romain Francois <francoiusromain@free.fr>
}
\keyword{ package }
